+++ components/layout_2020/formatting_contexts.rs
@@ -7,8 +7,7 @@ use crate::dom::NodeExt;
 use crate::dom_traversal::{Contents, NodeAndStyleInfo};
 use crate::flexbox::FlexContainer;
 use crate::flow::BlockFormattingContext;
-use crate::fragment_tree::BaseFragmentInfo;
-use crate::fragments::Fragment;
+use crate::fragment_tree::{BaseFragmentInfo, Fragment};
 use crate::positioned::PositioningContext;
 use crate::replaced::ReplacedContent;
 use crate::sizing::{self, ContentSizes};
