+++ components/layout_2020/fragment_tree/mod.rs
@@ -2,8 +2,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-mod base;
+mod base_fragment;
+mod box_fragment;
 mod containing_block;
+mod fragment;
+mod fragment_tree;
+mod hoisted_shared_fragment;
 
-pub(crate) use base::*;
+pub(crate) use base_fragment::*;
+pub(crate) use box_fragment::*;
 pub(crate) use containing_block::*;
+pub(crate) use fragment::*;
+pub use fragment_tree::*;
+pub(crate) use hoisted_shared_fragment::*;
